requirements.txt
Babel==2.10.3
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
pyserial==3.5
pytz==2022.4
schedule==1.1.0
tkcalendar==1.6.1
VimbaPython==1.2.0
